Added new Testmap (map004.tmx) and added the "Entity Layer " functionality.
The Entity layer is the layer, in which the entites are drawn. So first
all the ground layers, then all the layers above the "Entities" layer
are drawn. This allows the player to hide behind trees, for example.

I also added preperation libs and stuff to the classpath, so I can add
continuations back in.