Added "Fringe Layer" functionality. Tweaked continuation-AI.
The Fringe layer is a Tiled layer which makes it possible to have Tiles
behave like an Entity somehow. They get depth-sorted as well, and
therefore make it possible to be either drawn above an Entity (if the
Entity stands behind it) or below the Entity.

It also defines which layers generally draw above all entities and
below.

To make use of a Fringe layer one has to create a layer in the map and
call it "Fringe" (Upper or lower case doesn't matter). Every layer above
the Fringe layer will be generally drawn above the Entities, every layer
below, will be drawn below those.
To make Tiles from a Fringe layer appear in the game, one has to add a
"yoffset" property to all tiles one wants to use.
The "yoffset" property defines the relative y-position (relative to the
lower edge of the tile) where the tile meets the ground.
So if for example one wants to add a flower, which goes out of the
ground just in the middle of the tile, he has to use the yoffset 0.5
(m).
If one has the upper part of a flower, which should be connected to the
previously described lower part, then he has to use the yoffset -0.5 to
define the "meeting with the ground" at a lower level as the tile itself
is.

If you want to try it out yourself / play around with the values /
didn't get it yet, simply open up the map004.tmx with Tiled and per
right-click into the tileset look at the Properties which were defined
for the Flora used in the Fringe layer.